requirements.txt
flask
pandas
openpyxl
google-generativeai
werkzeug
